requirements.txt
bs4
requests
redis
google-search
google
wp-version-checker
mailchimp3
requests
firebase_admin
pika
flask
google-api-python-client
asyncio-nats-client
asyncio-nats-streaming